Some more MSVC warnings fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@28383 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/ili/ogrili1layer.cpp
@@ -428,7 +428,7 @@ void OGRILI1Layer::JoinSurfaceLayer( OGRILI1Layer* poSurfacePolyLayer, int nSurf
     poSurfacePolyLayer->ResetReading();
     while (OGRFeature *polyfeature = poSurfacePolyLayer->GetNextFeatureRef()) {
         GIntBig reftid = polyfeature->GetFID();
-        OGRFeature *feature = GetFeatureRef(reftid);
+        OGRFeature *feature = GetFeatureRef((int)reftid);
         if (feature) {
             feature->SetGeomField(nSurfaceFieldIndex, polyfeature->GetGeomFieldRef(0));
         } else {
